refact(Account): Add more precise Account interfaces
