+++ server/graphql/v2/object/Collective.js
@@ -1,17 +1,20 @@
 import { GraphQLObjectType } from 'graphql';
 
 import { types as collectiveTypes } from '../../../constants/collectives';
-import { Account, AccountFields, CollectiveAndFundFields } from '../interface/Account';
+import { Account, AccountFields } from '../interface/Account';
+import { AccountWithContributions, AccountWithContributionsFields } from '../interface/AccountWithContributions';
+import { AccountWithHost, AccountWithHostFields } from '../interface/AccountWithHost';
 
 export const Collective = new GraphQLObjectType({
   name: 'Collective',
   description: 'This represents a Collective account',
-  interfaces: () => [Account],
+  interfaces: () => [Account, AccountWithHost, AccountWithContributions],
   isTypeOf: collective => collective.type === collectiveTypes.COLLECTIVE,
   fields: () => {
     return {
       ...AccountFields,
-      ...CollectiveAndFundFields,
+      ...AccountWithHostFields,
+      ...AccountWithContributionsFields,
     };
   },
 });
