+++ server/graphql/v2/object/Fund.js
@@ -1,16 +1,19 @@
 import { GraphQLObjectType } from 'graphql';
 
-import { Account, AccountFields, CollectiveAndFundFields } from '../interface/Account';
+import { Account, AccountFields } from '../interface/Account';
+import { AccountWithContributions, AccountWithContributionsFields } from '../interface/AccountWithContributions';
+import { AccountWithHost, AccountWithHostFields } from '../interface/AccountWithHost';
 
 export const Fund = new GraphQLObjectType({
   name: 'Fund',
   description: 'This represents an Project account',
-  interfaces: () => [Account],
+  interfaces: () => [Account, AccountWithHost, AccountWithContributions],
   isTypeOf: collective => collective.type === 'FUND',
   fields: () => {
     return {
       ...AccountFields,
-      ...CollectiveAndFundFields,
+      ...AccountWithHostFields,
+      ...AccountWithContributionsFields,
     };
   },
 });
